split too big file in its natural chunks; make sure we attempt to copy arangods if packaging crashes

+++ release_tester/arangodb/installers/base.py
@@ -17,262 +17,14 @@ import yaml
 import psutil
 
 from arangodb.async_client import ArangoCLIprogressiveTimeoutExecutor, make_default_params
+from arangodb.installers.binary_description import BinaryDescription, FILE_EXTENSION, FILE_PIDS
 from arangodb.installers import InstallerConfig
 from arangodb.instance import ArangodInstance
 from tools.asciiprint import print_progress as progress
 from allure_commons._allure import attach
 from reporting.reporting_utils import step
 
-FILE_PIDS = []
-
-
 IS_WINDOWS = platform.win32_ver()[0]
-FILE_EXTENSION = ""
-if IS_WINDOWS:
-    FILE_EXTENSION = ".exe"
-
-IS_MAC = False
-if platform.mac_ver()[0]:
-    IS_MAC = True
-
-PROP_NAMES = {
-    'Comments': '@binary_version',
-    'InternalName': '@binary',
-    'ProductName': '@binary',
-    'CompanyName': '@company_name',
-    'LegalCopyright': '[ArangoDB GmbH ]*\(C\) Copyright 20.*',
-    'ProductVersion': '@windows_version',
-    'FileDescription': '@friendly_name',
-    'LegalTrademarks': None,
-    'PrivateBuild': None,
-    'FileVersion': '@windows_version',
-    'OriginalFilename': '@binary',
-    'SpecialBuild': None
-}
-## helper classes
-class BinaryDescription:
-    """describe the availability of an arangodb binary and its properties"""
-
-    def __init__(self, path, name, friendly, enter, strip, vmin, vmax, sym, binary_type):
-        # pylint: disable=too-many-arguments disable=too-many-instance-attributes
-        self.path = path / (name + FILE_EXTENSION)
-        self.enterprise = enter
-        self.stripped = strip
-        self.version_min =  semver.VersionInfo.parse(vmin)
-        self.version_max =  semver.VersionInfo.parse(vmax)
-        self.symlink = sym
-        self.binary_type = binary_type
-        self.name = name
-        self.friendly_name = friendly
-
-        for attribute in (
-            self.path,
-            self.enterprise,
-            self.stripped,
-            self.version_min,
-            self.version_max,
-            self.symlink,
-            self.binary_type,
-        ):
-            if attribute is None:
-                logging.error("one of the given args is null")
-                logging.error(str(self))
-                raise ValueError
-
-    def __repr__(self):
-        return """
-        path:        {0.path}
-        friendly name: {0.friendly_name}
-        enterprise:  {0.enterprise}
-        stripped:    {0.stripped}
-        version_min: {0.version_min}
-        version_max: {0.version_max}
-        symlink:     {0.symlink}
-        binary_type: {0.binary_type}
-        """.format(
-            self
-        )
-
-    def _validate_notarization(self, enterprise):
-        """check whether this binary is notarized"""
-        if not enterprise and self.enterprise:
-            return
-        if IS_MAC:
-            cmd = ["codesign", "--verify", "--verbose", str(self.path)]
-            check_strings = [b"valid on disk", b"satisfies its Designated Requirement"]
-            with psutil.Popen(cmd, bufsize=-1, stdout=subprocess.PIPE, stderr=subprocess.PIPE) as proc:
-                (_, codesign_str) = proc.communicate()
-                if proc.returncode:
-                    raise Exception("codesign exited nonzero " + str(cmd) + "\n" + str(codesign_str))
-                if codesign_str.find(check_strings[0]) < 0 or codesign_str.find(check_strings[1]) < 0:
-                    raise Exception("codesign didn't find signature: " + str(codesign_str))
-
-    def _binary(self, string):
-        """ should be our name """
-        return string == self.name
-
-    def _binary_version(self, version, string):
-        """ should be our name """
-        return string == f"{self.name} v{version.major}.{version.minor}" or string == "arangobench"
-
-    def _windows_version(self, version, version_string):
-        """ funny windows version should be there """
-        return f"{version.major}.{version.minor}.0.{version.patch}" == version_string
-
-    def _company_name(self, version, string):
-        """ may find company name """
-        # pylint: disable=chained-comparison
-        if string == "ArangoDB GmbH":
-            return True
-        if (version < semver.VersionInfo.parse("3.10.12") or
-            (version > semver.VersionInfo.parse("3.11.0") and version < semver.VersionInfo.parse("3.11.5") )):
-            return string is None or string == ""
-        return False
-
-    def _friendly_name(self, string):
-        """ should find friendly name in string """
-        return re.match(self.friendly_name, string) is not None
-
-    def _validate_windows_attributes(self, version, enterprise):
-        """ validate the windows binary header fields """
-        if not enterprise and self.enterprise:
-            return
-        if IS_WINDOWS and self.binary_type != "go": # GT-540: remove type comparison, alter versions.
-            # pylint: disable=import-outside-toplevel
-            from win32api import GetFileVersionInfo
-            language, codepage = GetFileVersionInfo(str(self.path), '\\VarFileInfo\\Translation')[0]
-            for windows_field, hook in PROP_NAMES.items():
-                exstr = ''
-                check_ok = True
-                try:
-                    string_file_info = '\\StringFileInfo\\%04X%04X\\%s' % (language, codepage, windows_field)
-                    description = GetFileVersionInfo(str(self.path), string_file_info)
-                except Exception as ex:
-                    description = None
-                    exstr = str(ex)
-                if description is None:
-                    if hook is not None:
-                        raise Exception(f"{windows_field} in '{self.path}' expected: to be set to , {hook}, but is None - {exstr}")
-                elif hook  == '@binary':
-                    check_ok = self._binary(description)
-                elif hook == '@binary_version':
-                    check_ok = self._binary_version(version, description)
-                elif hook  == '@company_name':
-                    check_ok = self._company_name(version, description)
-                elif hook  == '@windows_version':
-                    check_ok = self._windows_version(version, description)
-                elif hook  == '@friendly_name':
-                    check_ok = self._friendly_name(description)
-                else:
-                    check_ok = re.match(hook, description) is not None
-                if not check_ok:
-                    raise Exception(f"{windows_field} in '{self.path}' expected: to be set to , '{hook}', but is {description} - '{exstr}'")
-
-    # pylint: disable=too-many-arguments
-    @step
-    def check_installed(self, version, enterprise, check_stripped, check_symlink, check_notarized):
-        """check all attributes of this file in reality"""
-        attach(str(self), "file info")
-        if version > self.version_max or version < self.version_min:
-            self.check_path(enterprise, False)
-            return
-        self._validate_windows_attributes(version, enterprise)
-        if check_notarized:
-            self._validate_notarization(enterprise)
-        self.check_path(enterprise)
-        if not enterprise and self.enterprise:
-            # checks do not need to continue in this case
-            return
-        if check_stripped:
-            self.check_stripped()
-        if check_symlink:
-            self.check_symlink()
-
-    def check_path(self, enterprise, in_version=True):
-        """check whether the file rightfully exists or not"""
-        is_there = self.path.is_file()
-        if enterprise and self.enterprise:
-            if not is_there and in_version:
-                raise Exception("Binary missing from enterprise package! " + str(self.path))
-        # file must not exist
-        if not enterprise and self.enterprise:
-            if is_there:
-                raise Exception("Enterprise binary found in community package! " + str(self.path))
-        elif not is_there:
-            raise Exception("binary was not found! " + str(self.path))
-
-    def check_stripped_mac(self):
-        """Checking stripped status of the arangod"""
-        time.sleep(1)
-        if self.binary_type == "c++":
-            # finding out the file size before stripped cmd invoked
-            befor_stripped_size = self.path.stat().st_size
-
-            to_file = Path("/tmp/test_whether_stripped")
-            shutil.copy(str(self.path), str(to_file))
-            # invoke the strip command on file_path
-            cmd = ["strip", str(to_file)]
-            with subprocess.Popen(cmd, bufsize=-1, stderr=subprocess.PIPE, stdin=subprocess.PIPE) as proc:
-                FILE_PIDS.append(str(proc.pid))
-                proc.communicate()
-                proc.wait()
-            # check the size of copied file after stripped
-            after_stripped_size = to_file.stat().st_size
-
-            # cleanup temporary file:
-            if to_file.is_file():
-                to_file.unlink(str(to_file))
-            else:
-                print("stripped file not found")
-
-            # checking both output size
-            return befor_stripped_size == after_stripped_size
-        # some go binaries are stripped, some not. We can't test it.
-        return self.stripped
-
-    def check_stripped_windows(self):
-        """check whether this file is stripped (or not)"""
-        output = magic.from_file(str(self.path))
-        if output.find("PE32+") < 0:
-            raise Exception(f"Strip chinging for file {str(self.path)} returned [{output}]")
-        return output.find("(stripped") >= 0
-
-    def check_stripped_linux(self):
-        """check whether this file is stripped (or not)"""
-        output = magic.from_file(str(self.path))
-        if output.find("ELF") < 0:
-            raise Exception(f"Strip checking for file {str(self.path)} returned [{output}]")
-        if output.find(", stripped") >= 0:
-            return True
-        if output.find(", not stripped") >= 0:
-            return False
-        raise Exception(f"Strip checking: parse error for file '{str(self.path)}', unparseable output:  [{output}]")
-
-    @step
-    def check_stripped(self):
-        """check whether this file is stripped (or not)"""
-        is_stripped = True
-        if IS_MAC:
-            print("")
-            # is_stripped = self.check_stripped_mac()
-        elif IS_WINDOWS:
-            is_stripped = self.check_stripped_windows()
-        else:
-            is_stripped = self.check_stripped_linux()
-            if not is_stripped and self.stripped:
-
-                raise Exception("expected " + str(self.path) + " to be stripped, but it is not stripped")
-
-            if is_stripped and not self.stripped:
-                raise Exception("expected " + str(self.path) + " not to be stripped, but it is stripped")
-
-    @step
-    def check_symlink(self):
-        """check whether the file exists and is a symlink (if)"""
-        for link in self.symlink:
-            if not link.is_symlink():
-                raise Exception("{0} is not a symlink".format(str(link)))
-
 
 ### main class
 # pylint: disable=attribute-defined-outside-init disable=too-many-public-methods disable=too-many-instance-attributes
@@ -330,7 +82,8 @@ class InstallerBase(ABC):
         eps = "E_" if self.cfg.enterprise else ""
         self.backup_arangod_name = self.cfg.base_test_dir / f"arangod_{eps}_{self.cfg.version}{FILE_EXTENSION}"
         arangod_src = self.cfg.real_sbin_dir / f"arangod{FILE_EXTENSION}"
-        if not str(self.backup_arangod_name) in self.arangods:
+        if (not str(self.backup_arangod_name) in self.arangods and
+            arangod_src.exists()):
             self.arangods.append(self.backup_arangod_name)
             print("copying " + str(arangod_src) + " to " + str(self.backup_arangod_name))
             shutil.copy(str(arangod_src), str(self.backup_arangod_name))
@@ -823,6 +576,10 @@ class InstallerBase(ABC):
                 success = False
         return success
 
+    def get_log_dir(self):
+        """ get the directory the installer will create logs in """
+        return self.cfg.install_prefix / self.cfg.log_dir
+
     def set_system_instance(self):
         """
         set an instance representing the system service launched by packages
@@ -832,7 +589,7 @@ class InstallerBase(ABC):
             port="8529",
             localhost=self.cfg.localhost,
             publicip=self.cfg.publicip,
-            basedir=(self.cfg.install_prefix / self.cfg.log_dir),
+            basedir=(self.get_log_dir()),
             passvoid=self.cfg.passvoid,
             ssl=False,
             version=self.cfg.version,
